# Use unmarshal_unsafe option to reduce allocs on Protobuf unmarshalling