… client.pb.go
… client_vtproto.pb.go
… go.sum