Before:

```
BenchmarkReplyProtobufUnmarshal-12            	 6853702	       163.2 ns/op	     277 B/op	       3 allocs/op
BenchmarkReplyProtobufUnmarshalParallel-12    	13072430	        89.92 ns/op	     277 B/op	       3 allocs/op
```

After:

```
BenchmarkReplyProtobufUnmarshal-12            	 8410488	       141.0 ns/op	     272 B/op	       2 allocs/op
BenchmarkReplyProtobufUnmarshalParallel-12    	14639738	        81.16 ns/op	     272 B/op	       2 allocs/op
```